tweaking latest merge from upstream

+++ sysproduction/interactive_order_stack.py
@@ -223,14 +223,12 @@ def get_broker_order_details_for_balance_trade(data: dataBlob) -> brokerOrder:
         (
             instrument_code,
             contract_date_yyyy_mm,
-            # TODO AG position int -> float
             fill_qty,
         ) = get_futures_contract_and_qty_to_close_position(data)
     else:
         (
             instrument_code,
             contract_date_yyyy_mm,
-            # TODO AG position int -> float
             fill_qty,
         ) = manually_get_futures_contract_and_qty(data)
 
@@ -295,8 +293,12 @@ def manually_get_futures_contract_and_qty(data: dataBlob) -> Tuple[str, str, int
 
     print("Actual contract expiry is %s" % str(actual_contract_date))
 
+    # TODO AG position int -> float
+    # fill_qty = get_input_from_user_and_convert_to_type(
+    #     "Quantity ", type_expected=int, allow_default=False
+    # )
     fill_qty = get_input_from_user_and_convert_to_type(
-        "Quantity ", type_expected=int, allow_default=False
+        "Quantity ", type_expected=float, allow_default=False
     )
 
     return instrument_code, contract_date_yyyy_mm, fill_qty
@@ -323,7 +325,9 @@ def get_futures_contract_and_qty_to_close_position(
     relevant_row = contract_positions[position_index]
     instrument_code = relevant_row.instrument_code
     contract_date_yyyy_mm = relevant_row.date_str
-    fill_qty = int(-relevant_row.position)
+    # fill_qty = int(-relevant_row.position)
+    # TODO AG position int -> float
+    fill_qty = -relevant_row.position
 
     return instrument_code, contract_date_yyyy_mm, fill_qty
